% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplier.R
\name{multiplier}
\alias{multiplier}
\title{A function for multiplying two numbers}
\usage{
multiplier(a, b)
}
\arguments{
\item{a}{A real number}

\item{b}{a real number}
}
\value{
the product of \code{a} and \code{b}
}
\description{
A baby function to illustrate how to document stuff in R
It is soley a function to multiply two numbers
}
\examples{
multiplier(2, 5)
multiplier(4, 5)
}
